+++ tests/exceptions/test_exceptions.py
@@ -1,5 +1,6 @@
 """Unit tests for the Exceptions module."""
 import os
+import tempfile
 
 import pytest
 
@@ -7,10 +8,16 @@ from statick_tool.exceptions import Exceptions
 from statick_tool.issue import Issue
 from statick_tool.package import Package
 
+try:
+    from tempfile import TemporaryDirectory
+except:  # pylint: disable=bare-except # noqa: E722 # NOLINT
+    from backports.tempfile import (  # pylint: disable=wrong-import-order
+        TemporaryDirectory,
+    )
+
 
 def test_exceptions_init_valid():
-    """
-    Test that the Exceptions module initializes correctly.
+    """Test that the Exceptions module initializes correctly.
 
     Expected result: exceptions.exceptions is initialized.
     """
@@ -21,8 +28,7 @@ def test_exceptions_init_valid():
 
 
 def test_exceptions_init_nonexistent():
-    """
-    Test that the Exceptions module throws an OSError if a bad path is given.
+    """Test that the Exceptions module throws an OSError if a bad path is given.
 
     Expected result: OSError thrown.
     """
@@ -33,8 +39,7 @@ def test_exceptions_init_nonexistent():
 
 
 def test_exceptions_file_empty_string():
-    """
-    Test for when a Exceptions is initialized with an empty string.
+    """Test for when a Exceptions is initialized with an empty string.
 
     Expected result: ValueError is thrown
     """
@@ -43,8 +48,7 @@ def test_exceptions_file_empty_string():
 
 
 def test_exceptions_file_invalid_yaml():
-    """
-    Test for when a Exceptions is initialized with an invalid yaml file.
+    """Test for when a Exceptions is initialized with an invalid yaml file.
 
     Expected result: ValueError is thrown
     """
@@ -53,8 +57,7 @@ def test_exceptions_file_invalid_yaml():
 
 
 def test_filter_file_exceptions_early():
-    """
-    Test that filter_file_exceptions_early excludes files.
+    """Test that filter_file_exceptions_early excludes files.
 
     Expected result: Empty files list.
     """
@@ -74,8 +77,7 @@ def test_filter_file_exceptions_early():
 
 
 def test_filter_file_exceptions_early_onlyall():
-    """
-    Test that filter_file_exceptions_early only uses exceptions with tools=all.
+    """Test that filter_file_exceptions_early only uses exceptions with tools=all.
 
     Expected result: No change to the files list
     """
@@ -92,8 +94,7 @@ def test_filter_file_exceptions_early_onlyall():
 
 
 def test_filter_file_exceptions_early_dupes():
-    """
-    Test that filter_file_exceptions_early excludes duplicated files.
+    """Test that filter_file_exceptions_early excludes duplicated files.
 
     I have no idea why one might have duplicate files, but might as well test it!
     Expected result: Empty file list.
@@ -113,9 +114,42 @@ def test_filter_file_exceptions_early_dupes():
     assert not filtered_files
 
 
-def test_global_exceptions():
+def test_ignore_packages():
     """
-    Test that global exceptions are found.
+    Test that ignored packages are read correctly.
+
+    Expected result: List of ignored packages matches configuration file.
+    """
+    # Look at file without "ignore_packages" key.
+    exceptions = Exceptions(
+        os.path.join(os.path.dirname(__file__), "early_exceptions.yaml")
+    )
+    expected = []
+    ignored_packages = exceptions.get_ignore_packages()
+
+    assert expected == ignored_packages
+
+    # Look at file with "ignore_packages" key but no packages specified.
+    exceptions = Exceptions(
+        os.path.join(os.path.dirname(__file__), "ignore_package_none_exceptions.yaml")
+    )
+    expected = []
+    ignored_packages = exceptions.get_ignore_packages()
+
+    assert expected == ignored_packages
+
+    # Look at file with "ignore_packages" key and packages specified.
+    exceptions = Exceptions(
+        os.path.join(os.path.dirname(__file__), "ignore_package_exceptions.yaml")
+    )
+    expected = ["package_a", "package_b"]
+    ignored_packages = exceptions.get_ignore_packages()
+
+    assert expected == ignored_packages
+
+
+def test_global_exceptions():
+    """Test that global exceptions are found.
 
     Expected result: one global exception each for file and message_regex.
     """
@@ -132,8 +166,7 @@ def test_global_exceptions():
 
 
 def test_package_exceptions():
-    """
-    Test that package exceptions are found.
+    """Test that package exceptions are found.
 
     Expected result: exceptions are found for both file and message_regex types
     """
@@ -150,8 +183,7 @@ def test_package_exceptions():
 
 
 def test_filter_issues():
-    """
-    Test that issues are filtered based on regex exceptions.
+    """Test that issues are filtered based on regex exceptions.
 
     Expected result: all issues are filtered and none are found
     """
@@ -177,8 +209,7 @@ def test_filter_issues():
 
 
 def test_filter_issues_empty_exceptions():
-    """
-    Test that issues are filtered when the exceptions file is empty.
+    """Test that issues are filtered when the exceptions file is empty.
 
     Expected result: one issue is found.
     """
@@ -204,8 +235,7 @@ def test_filter_issues_empty_exceptions():
 
 
 def test_filter_issues_globs():
-    """
-    Test that issues are filtered based on regex exceptions if it matches a glob.
+    """Test that issues are filtered based on regex exceptions if it matches a glob.
 
     Expected result: all issues are filtered out.
     """
@@ -231,8 +261,7 @@ def test_filter_issues_globs():
 
 
 def test_filter_issues_globs_wrong_file_pattern():
-    """
-    Test that issues are filtered based on regex exceptions if it matches a glob.
+    """Test that issues are filtered based on regex exceptions if it matches a glob.
 
     Expected result: no issues are filtered and one issue is found.
     """
@@ -258,8 +287,7 @@ def test_filter_issues_globs_wrong_file_pattern():
 
 
 def test_filter_issues_travis_build():
-    """
-    Test that issues on Travis CI are not filtered based on the filename prefix.
+    """Test that issues on Travis CI are not filtered based on the filename prefix.
 
     Expected result: all but one non-excepted issue is filtered
     """
@@ -285,8 +313,7 @@ def test_filter_issues_travis_build():
 
 
 def test_filter_issues_filename_abs_path():
-    """
-    Test that issues are filtered based on regex exceptions with absolute path.
+    """Test that issues are filtered based on regex exceptions with absolute path.
 
     Expected result: no issues found
     """
@@ -312,8 +339,7 @@ def test_filter_issues_filename_abs_path():
 
 
 def test_filter_issues_nolint():
-    """
-    Test that issues are filtered based on NOLINT comment.
+    """Test that issues are filtered based on NOLINT comment.
 
     Expected result: no issues found
     """
@@ -339,8 +365,7 @@ def test_filter_issues_nolint():
 
 
 def test_filter_issues_nolint_empty_log():
-    """
-    Test that NOLINT excpetions to issues do not fail with an empty issue log file.
+    """Test that NOLINT excpetions to issues do not fail with an empty issue log file.
 
     Expected result: same number of original issues in filtered issues
     """
@@ -362,9 +387,67 @@ def test_filter_issues_nolint_empty_log():
     assert len(issues) == len(filtered_issues)
 
 
-def test_filter_issues_nolint_not_abs_path():
+def test_filter_issues_nolint_unicode_decode_error():
     """
-    Test that issues are not filtered based on NOLINT comment when not absolute path.
+    Test that excpetions do not fail with a file known to cause UnicodeDecodeError.
+
+    Example file that causes UnicodeDecodeError is from
+    https://github.com/PointCloudLibrary/blog.
+
+    Expected result: same number of original issues in filtered issues
+    """
+    exceptions = Exceptions(
+        os.path.join(os.path.dirname(__file__), "valid_exceptions.yaml")
+    )
+
+    filename = (
+        os.path.join(os.path.dirname(__file__), "unicode_decode_error_package")
+        + "/status.rst"
+    )
+    line_number = "0"
+    tool = "dummy_tool"
+    issue_type = "dummy_issue_type"
+    severity = "0"
+    message = "dummy_message"
+    tool_issue = Issue(filename, line_number, tool, issue_type, severity, message, None)
+    issues = {}
+    issues["dummy_tool"] = [tool_issue]
+
+    filtered_issues = exceptions.filter_nolint(issues)
+    assert len(issues) == len(filtered_issues)
+
+
+def test_filter_issues_nolint_file_not_found_error():
+    """
+    Test that excpetions do not fail with a file that does not exist.
+
+    Expected result: same number of original issues in filtered issues
+    """
+    exceptions = Exceptions(
+        os.path.join(os.path.dirname(__file__), "valid_exceptions.yaml")
+    )
+
+    with TemporaryDirectory() as tmp_dir:
+        # Make a temporary executable
+        with tempfile.NamedTemporaryFile(dir=tmp_dir) as tmp_file:
+            filename = tmp_file.name
+            line_number = "0"
+            tool = "dummy_tool"
+            issue_type = "dummy_issue_type"
+            severity = "0"
+            message = "dummy_message"
+            tool_issue = Issue(
+                filename, line_number, tool, issue_type, severity, message, None
+            )
+            issues = {}
+            issues["dummy_tool"] = [tool_issue]
+
+    filtered_issues = exceptions.filter_nolint(issues)
+    assert len(issues) == len(filtered_issues)
+
+
+def test_filter_issues_nolint_not_abs_path():
+    """Test that issues are not filtered based on NOLINT comment when not absolute path.
 
     Expected result: one issue found
     """
@@ -390,8 +473,8 @@ def test_filter_issues_nolint_not_abs_path():
 
 
 def test_filter_issues_wildcard_exceptions():
-    """
-    Test that issues are found even when exceptions with wildcards for regex are used.
+    """Test that issues are found even when exceptions with wildcards for regex are
+    used.
 
     Expected result: one issue found
     """