+++ tests/plugins/tool/groovylint_tool_plugin/test_groovylint_tool_plugin.py
@@ -127,8 +127,8 @@ def test_groovylint_tool_plugin_parse_invalid():
 
 @mock.patch("statick_tool.plugins.tool.groovylint_tool_plugin.subprocess.check_output")
 def test_groovylint_tool_plugin_scan_calledprocesserror(mock_subprocess_check_output):
-    """
-    Test what happens when a CalledProcessError is raised (usually means groovylint hit an error).
+    """Test what happens when a CalledProcessError is raised (usually means groovylint
+    hit an error).
 
     Expected result: no issues found
     """
@@ -154,8 +154,8 @@ def test_groovylint_tool_plugin_scan_calledprocesserror(mock_subprocess_check_ou
 
 @mock.patch("statick_tool.plugins.tool.groovylint_tool_plugin.subprocess.check_output")
 def test_groovylint_tool_plugin_scan_oserror(mock_subprocess_check_output):
-    """
-    Test what happens when an OSError is raised (usually means groovylint doesn't exist).
+    """Test what happens when an OSError is raised (usually means groovylint doesn't
+    exist).
 
     Expected result: issues is None
     """