+++ tests/plugins/tool/make_tool_plugin/test_make_tool_plugin.py
@@ -196,8 +196,8 @@ def test_make_tool_plugin_parse_warnings_mapping():
 
 @mock.patch("statick_tool.plugins.tool.make_tool_plugin.subprocess.check_output")
 def test_make_tool_plugin_scan_calledprocesserror(mock_subprocess_check_output):
-    """
-    Test what happens when a CalledProcessError is raised (usually means make hit an error).
+    """Test what happens when a CalledProcessError is raised (usually means make hit an
+    error).
 
     Expected result: issues is None
     """
@@ -215,8 +215,7 @@ def test_make_tool_plugin_scan_calledprocesserror(mock_subprocess_check_output):
 
 @mock.patch("statick_tool.plugins.tool.make_tool_plugin.subprocess.check_output")
 def test_make_tool_plugin_scan_oserror(mock_subprocess_check_output):
-    """
-    Test what happens when an OSError is raised (usually means make doesn't exist).
+    """Test what happens when an OSError is raised (usually means make doesn't exist).
 
     Expected result: issues is None
     """