+++ tests/plugins/tool/mypy_tool_plugin/test_mypy_tool_plugin.py
@@ -111,8 +111,7 @@ def test_mypy_tool_plugin_parse_invalid():
 
 @mock.patch("statick_tool.plugins.tool.mypy_tool_plugin.subprocess.check_output")
 def test_mypy_tool_plugin_scan_oserror(mock_subprocess_check_output):
-    """
-    Test what happens when an OSError is raised (usually means mypy doesn't exist).
+    """Test what happens when an OSError is raised (usually means mypy doesn't exist).
 
     Expected result: issues is None
     """
@@ -130,8 +129,8 @@ def test_mypy_tool_plugin_scan_oserror(mock_subprocess_check_output):
 
 @mock.patch("statick_tool.plugins.tool.mypy_tool_plugin.subprocess.check_output")
 def test_mypy_tool_plugin_scan_calledprocesserror(mock_subprocess_check_output):
-    """
-    Test what happens when a CalledProcessError is raised (usually means mypy hit an error).
+    """Test what happens when a CalledProcessError is raised (usually means mypy hit an
+    error).
 
     Expected result: issues is None
     """