+++ tests/plugins/tool/pydocstyle_tool_plugin/test_pydocstyle_tool_plugin.py
@@ -94,8 +94,8 @@ def test_pydocstyle_tool_plugin_parse_invalid():
 
 @mock.patch("statick_tool.plugins.tool.pydocstyle_tool_plugin.subprocess.check_output")
 def test_pydocstyle_tool_plugin_scan_calledprocesserror(mock_subprocess_check_output):
-    """
-    Test what happens when a CalledProcessError is raised (usually means tool hit an error).
+    """Test what happens when a CalledProcessError is raised (usually means tool hit an
+    error).
 
     Expected result: issues is None
     """
@@ -121,8 +121,7 @@ def test_pydocstyle_tool_plugin_scan_calledprocesserror(mock_subprocess_check_ou
 
 @mock.patch("statick_tool.plugins.tool.pydocstyle_tool_plugin.subprocess.check_output")
 def test_pydocstyle_tool_plugin_scan_oserror(mock_subprocess_check_output):
-    """
-    Test what happens when an OSError is raised (usually means tool doesn't exist).
+    """Test what happens when an OSError is raised (usually means tool doesn't exist).
 
     Expected result: issues is None
     """